Correct tests for move of summary_by_vol
